Make state PLAYING on game start
